+++ CHANGES.md
@@ -5,9 +5,11 @@ All notable changes to this project will be documented in this file.
 
  [v0.3.0]  - unreleased
 ---------------------------------
-* add rtp_forward interface to the admin API of videoroom plugin
-
-
+* add rtp_forward operations to the admin API of videoroom plugin
+* Add support for VP9 and H.264 profile negotiation for videoroom and echotest plugin
+* Added support for multichannel Opus audio (surround) for videoroom
+* Added VideoRoom option to only allow admins to change the recording state
+* Enable / disable recording while conference is in progress for videoroom
 
  [v0.2.0]  - 2020-05-10
 ---------------------------------